+++ backapp/internal/router/router.go
@@ -3,9 +3,9 @@ package router
 import (
 	"backapp/internal/config"
 	"backapp/internal/handler"
+	"backapp/internal/middleware"
 	"backapp/internal/repository"
 	"database/sql"
-	"net/http"
 
 	"github.com/gin-gonic/gin"
 )
@@ -15,23 +15,7 @@ func SetupRouter(db *sql.DB, cfg *config.Config) *gin.Engine {
 	router := gin.Default()
 
 	// CORS middleware
-	router.Use(func(c *gin.Context) {
-		origin := c.Request.Header.Get("Origin")
-		if origin == "http://localhost:3300" || origin == "http://localhost:3000" {
-			c.Header("Access-Control-Allow-Origin", origin)
-		}
-		c.Header("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
-		c.Header("Access-Control-Allow-Headers", "Origin, Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, Cookie")
-		c.Header("Access-Control-Allow-Credentials", "true")
-		c.Header("Access-Control-Max-Age", "86400")
-
-		if c.Request.Method == "OPTIONS" {
-			c.AbortWithStatus(http.StatusNoContent)
-			return
-		}
-
-		c.Next()
-	})
+	router.Use(middleware.CORSMiddleware())
 
 	userRepo := repository.NewUserRepository(db)
 	authHandler := handler.NewAuthHandler(cfg, userRepo)
@@ -57,13 +41,13 @@ func SetupRouter(db *sql.DB, cfg *config.Config) *gin.Engine {
 				google.GET("/login", authHandler.GoogleLogin)
 				google.GET("/callback", authHandler.GoogleCallback)
 			}
-			auth.GET("/user", authHandler.AuthMiddleware(), authHandler.GetUser)
+			auth.GET("/user", middleware.AuthMiddleware(userRepo), authHandler.GetUser)
 			auth.GET("/logout", authHandler.Logout)
 		}
 
 		user := api.Group("/user")
 		{
-			user.Use(authHandler.AuthMiddleware())
+			user.Use(middleware.AuthMiddleware(userRepo))
 			user.PUT("/profile", authHandler.UpdateProfile)
 		}
 	}
